requirements.txt
pyautogui
opencv-python
pillow
pyinstaller
mss
python-dotenv